Fix horse races based on years
TJK's data says that a horse's age is calculated based on the first day
of year. At the start of each year the horses' ages are re-calculated.
So the race results from previous years contain age as 1 less year, for
example. To fix this, we add the amount of years between the race date
and now to the age when querying the horse database.